requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
pytz==2023.3
gTTS==2.4.0
